requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pgvector==0.3.6

# Validation / Env
pydantic==2.9.2
python-dotenv==1.0.1

# Security
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# HTTP Client (if needed)
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.23.8

# Linting & Formatting (optional but industry standard)
black==24.8.0
isort==5.13.2
flake8==7.1.1

# RAG related
groq==0.9.0
sentence-transformers==3.0.1
faiss-cpu==1.9.0
